refactor: change package rpc to handler
