+++ pkg/handler/model.go
@@ -1,4 +1,4 @@
-package rpc
+package handler
 
 import (
 	"archive/zip"
